# docs/swagger.yaml
basePath: /
definitions:
  request.AddCarRequest:
    properties:
      carId:
        type: string
      garageId:
        type: string
    type: object
...
  request.GarageRequest:
    properties:
      carId:
        type: string
    type: object
...
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is the doumentation for drag racing game
  title: Drag Racing
  version: "2.0"
paths:
...
  /garage/add-car:
    post:
      consumes:
      - application/json
      description: Add a car to a player's garage
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add Car Request
        in: body
        name: addCarRequest
        required: true
        schema:
          $ref: '#/definitions/request.AddCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Car added to garage successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Add Car to Garage
      tags:
      - Garage
  /garage/buy:
    post:
      consumes:
      - application/json
      description: Buy a garage for a player
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Buy Garage Request
        in: body
        name: buyRequest
        required: true
        schema:
          $ref: '#/definitions/request.GarageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Garage bought successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Garage not found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      security:
      - ApiKeyAuth: []
      summary: Buy Garage
      tags:
      - Garage
  /garage/get:
    get:
      consumes:
      - application/json
      description: Retrieve the list of garages owned by a player
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data fetched successfully
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get Player Garage List
      tags:
      - Garage
  /garage/upgrade:
    put:
      consumes:
      - application/json
      description: Upgrade a player's garage to the next level
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.GarageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Garage upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade Garage
      tags:
      - Garage
  /garages/get-all:
    get:
      consumes:
      - application/json
      description: Retrieve the list of all garages
      produces:
      - application/json
      responses:
        "200":
          description: Garage list fetched successfully
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get All Garage List
      tags:
      - Garage
...
swagger: "2.0"